requirements.txt
aiohttp==3.6.3
APScheduler==3.7.0
async-timeout==3.0.1
black==21.7b0
cryptography==3.4.8
decorator==5.1.0
discord-ext-menus==1.1
DiscordUtils==1.3.4
enhanced-dpy
pretty-errors==1.2.24
python-dotenv==0.19.0
pytz==2021.1
qrcode==7.3.1
regex==2021.8.28
requests==2.26.0
SQLAlchemy==1.4.25
sqlalchemy2-stubs==0.0.2a17
sqlmodel==0.0.4
sqlparse==0.4.2
web3==5.19.0
pycoingecko==3.1.0
